improve pdb status check
Prior to this commit, we were checking the pdb
version endpoint. It seems that in some cases
that endpoint will return a valid result before
pdb is actually available to process queries.
As a result, we think the service is up, continue
the test, try to do our final agent run, and
fail becuase puppetdb isn't ready.

This commit changes our status check to make a
real request to the pdb query endpoint as
the test for determining if the service is up.
All retry and wait limits remain the same.